fix: broker gaps - constants, QueueInfo types, topic value
- Add BROKER_INMEMORY, BROKER_RABBITMQ, TOPIC_JOB, TOPIC_JOB_COMPLETED, TOPIC_JOB_FAILED constants
- Add TOPIC_SCHEDULED_JOB constant to tork/broker/mod.rs (was only in scattered locations)
- Add Provider type alias
- Change QueueInfo fields from i64 to i32 to match Go's int
- Fix TOPIC_SCHEDULED_JOB value from 'scheduled-hyphen' to 'scheduled.dot' to match Go
- Update test_constants_match_go to expect 'scheduled.job'

GAP status:
- GAP 1 (constants): IMPLEMENTED
- GAP 2 (Provider type): IMPLEMENTED
- GAP 3 (context params): NOT IMPLEMENTED (breaking change - noted but deferred)
- GAP 4 (QueueInfo i32): IMPLEMENTED
- GAP 5 (TOPIC_SCHEDULED_JOB value): IMPLEMENTED
- GAP 6 (topic constants in broker module): IMPLEMENTED (constants now in tork/broker/mod.rs)

+++ broker/inmemory.rs
@@ -443,9 +443,9 @@ impl Broker for InMemoryBroker {
             let result: Vec<QueueInfo> = queues
                 .iter()
                 .map(|(name, q)| {
-                    let sub_count = q.subs.lock().map_or(0, |s| s.len()) as i64;
-                    let unacked = q.unacked.load(Ordering::SeqCst) as i64;
-                    let size = q.size.load(Ordering::SeqCst) as i64;
+                    let sub_count = q.subs.lock().map_or(0, |s| s.len()) as i32;
+                    let unacked = q.unacked.load(Ordering::SeqCst) as i32;
+                    let size = q.size.load(Ordering::SeqCst) as i32;
                     QueueInfo {
                         name: name.clone(),
                         size,
@@ -466,9 +466,9 @@ impl Broker for InMemoryBroker {
             queues
                 .get(&qname)
                 .map(|q| {
-                    let sub_count = q.subs.lock().map_or(0, |s| s.len()) as i64;
-                    let unacked = q.unacked.load(Ordering::SeqCst) as i64;
-                    let size = q.size.load(Ordering::SeqCst) as i64;
+                    let sub_count = q.subs.lock().map_or(0, |s| s.len()) as i32;
+                    let unacked = q.unacked.load(Ordering::SeqCst) as i32;
+                    let size = q.size.load(Ordering::SeqCst) as i32;
                     QueueInfo {
                         name: qname,
                         size,
